+++ sdk/highlight-run/CHANGELOG.md
@@ -1,5 +1,19 @@
 # highlight.run
 
+## [9.23.0](https://github.com/launchdarkly/observability-sdk/compare/highlight.run-9.22.5...highlight.run-9.23.0) (2025-11-12)
+
+
+### Features
+
+* disable client-side LD metric forwarding ([#286](https://github.com/launchdarkly/observability-sdk/issues/286)) ([648ad26](https://github.com/launchdarkly/observability-sdk/commit/648ad263d34442a9703102144455304bbb0be825))
+
+## [9.22.5](https://github.com/launchdarkly/observability-sdk/compare/highlight.run-9.22.4...highlight.run-9.22.5) (2025-11-10)
+
+
+### Bug Fixes
+
+* correctly forward error payload to launchdarkly ([#285](https://github.com/launchdarkly/observability-sdk/issues/285)) ([86fc7aa](https://github.com/launchdarkly/observability-sdk/commit/86fc7aa03768f5cba566c2208cc16dd558d3d17b))
+
 ## [9.22.4](https://github.com/launchdarkly/observability-sdk/compare/highlight.run-9.22.3...highlight.run-9.22.4) (2025-10-30)